+++ src/main/java/it/gov/pagopa/pu/classification/connector/debtposition/config/DebtPositionApiClientConfig.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.classification.connector.debtposition.config;
 
-import it.gov.pagopa.pu.classification.config.ApiClientConfig;
+import it.gov.pagopa.pu.classification.config.rest.ApiClientConfig;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
