+++ src/main/java/it/gov/pagopa/pu/classification/connector/processexecutions/config/ProcessExecutionsApiClientConfig.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.classification.connector.processexecutions.config;
 
-import it.gov.pagopa.pu.classification.config.ApiClientConfig;
+import it.gov.pagopa.pu.classification.config.rest.ApiClientConfig;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
