+++ src/test/java/it/gov/pagopa/pu/classification/util/HttpUtilsTest.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.classification.util;
 
-import it.gov.pagopa.pu.classification.config.HttpClientConfig;
+import it.gov.pagopa.pu.classification.config.rest.HttpClientConfig;
 import org.apache.hc.client5.http.HttpRoute;
 import org.apache.hc.client5.http.classic.HttpClient;
 import org.apache.hc.client5.http.config.ConnectionConfig;
